+++ api/v2/api.go
@@ -138,10 +138,11 @@ func (api *API) getStatusHandler(params general_ops.GetStatusParams) middleware.
 	api.mtx.RLock()
 	defer api.mtx.RUnlock()
 
-	name := api.peer.Name()
-	status := api.peer.Status()
 	original := api.alertmanagerConfig.String()
 	uptime := strfmt.DateTime(api.uptime)
+
+	status := open_api_models.ClusterStatusStatusDisabled
+
 	resp := open_api_models.AlertmanagerStatus{
 		Uptime: &uptime,
 		VersionInfo: &open_api_models.VersionInfo{
@@ -156,18 +157,28 @@ func (api *API) getStatusHandler(params general_ops.GetStatusParams) middleware.
 			Original: &original,
 		},
 		Cluster: &open_api_models.ClusterStatus{
-			Name:   &name,
 			Status: &status,
-			Peers:  []*open_api_models.PeerStatus{},
 		},
 	}
 
-	for _, n := range api.peer.Peers() {
-		address := n.Address()
-		resp.Cluster.Peers = append(resp.Cluster.Peers, &open_api_models.PeerStatus{
-			Name:    &n.Name,
-			Address: &address,
-		})
+	// If alertmanager cluster feature is disabled, then api.peers == nil.
+	if api.peer != nil {
+		status := api.peer.Status()
+
+		peers := []*open_api_models.PeerStatus{}
+		for _, n := range api.peer.Peers() {
+			address := n.Address()
+			peers = append(peers, &open_api_models.PeerStatus{
+				Name:    &n.Name,
+				Address: &address,
+			})
+		}
+
+		resp.Cluster = &open_api_models.ClusterStatus{
+			Name:   api.peer.Name(),
+			Status: &status,
+			Peers:  peers,
+		}
 	}
 
 	return general_ops.NewGetStatusOK().WithPayload(&resp)