+++ api/v2/models/cluster_status.go
@@ -6,6 +6,7 @@ package models
 // Editing this file might prove futile when you re-run the swagger generate command
 
 import (
+	"encoding/json"
 	"strconv"
 
 	strfmt "github.com/go-openapi/strfmt"
@@ -20,26 +21,21 @@ import (
 type ClusterStatus struct {
 
 	// name
-	// Required: true
-	Name *string `json:"name"`
+	Name string `json:"name,omitempty"`
 
 	// peers
-	// Required: true
 	// Minimum: 0
 	Peers []*PeerStatus `json:"peers"`
 
 	// status
 	// Required: true
+	// Enum: [ready settling disabled]
 	Status *string `json:"status"`
 }
 
 // Validate validates this cluster status
 func (m *ClusterStatus) Validate(formats strfmt.Registry) error {
 	var res []error
-
-	if err := m.validateName(formats); err != nil {
-		res = append(res, err)
-	}
 
 	if err := m.validatePeers(formats); err != nil {
 		res = append(res, err)
@@ -55,19 +51,10 @@ func (m *ClusterStatus) Validate(formats strfmt.Registry) error {
 	return nil
 }
 
-func (m *ClusterStatus) validateName(formats strfmt.Registry) error {
-
-	if err := validate.Required("name", "body", m.Name); err != nil {
-		return err
-	}
-
-	return nil
-}
-
 func (m *ClusterStatus) validatePeers(formats strfmt.Registry) error {
 
-	if err := validate.Required("peers", "body", m.Peers); err != nil {
-		return err
+	if swag.IsZero(m.Peers) { // not required
+		return nil
 	}
 
 	for i := 0; i < len(m.Peers); i++ {
@@ -89,9 +76,46 @@ func (m *ClusterStatus) validatePeers(formats strfmt.Registry) error {
 	return nil
 }
 
+var clusterStatusTypeStatusPropEnum []interface{}
+
+func init() {
+	var res []string
+	if err := json.Unmarshal([]byte(`["ready","settling","disabled"]`), &res); err != nil {
+		panic(err)
+	}
+	for _, v := range res {
+		clusterStatusTypeStatusPropEnum = append(clusterStatusTypeStatusPropEnum, v)
+	}
+}
+
+const (
+
+	// ClusterStatusStatusReady captures enum value "ready"
+	ClusterStatusStatusReady string = "ready"
+
+	// ClusterStatusStatusSettling captures enum value "settling"
+	ClusterStatusStatusSettling string = "settling"
+
+	// ClusterStatusStatusDisabled captures enum value "disabled"
+	ClusterStatusStatusDisabled string = "disabled"
+)
+
+// prop value enum
+func (m *ClusterStatus) validateStatusEnum(path, location string, value string) error {
+	if err := validate.Enum(path, location, value, clusterStatusTypeStatusPropEnum); err != nil {
+		return err
+	}
+	return nil
+}
+
 func (m *ClusterStatus) validateStatus(formats strfmt.Registry) error {
 
 	if err := validate.Required("status", "body", m.Status); err != nil {
+		return err
+	}
+
+	// value enum
+	if err := m.validateStatusEnum("status", "body", *m.Status); err != nil {
 		return err
 	}